+++ peer-to-peer/client.py
@@ -1,83 +1,113 @@
-
 from twisted.internet.protocol import DatagramProtocol
 from twisted.internet import reactor
 from random import randint
-import time
-
 
 class Client(DatagramProtocol):
 
     def __init__(self, port):
         host = "127.0.0.1"
         self.id = host, port
+        
         self.addressToSendMsg = None
-        self.clientList = []
-        print('my address:', self.id, " Ready to send and recieve")
-        time.sleep(1)
-
+        self.addressLastSent = None
+        self.addressLastReceived = None
+        
+        self.peers = dict()
+                
+        print("P2P Messanger Started @", self.id)
+        print("Ready to send and Recieve \n\n")
+    
+    #runs when reactor.run() is called
     def startProtocol(self):
-        self.transport.write("ready".encode('utf_8'), self.id)
-
-    def messageClients(self):
-        print("Select option 1 or 2 to send message")
-        while True:
-            print("Select option 1 or 2 to send message")
-            time.sleep(0.5)
-            print("1 - message to peer clients")
-            print("2 - message to new client")
-            print("3 - Exit from the application")
-            try:
-                option = int(input("Enter your option: "))
-                if (option not in (1, 2, 3)):
-                    print("Invalid input")
-                elif option == 1:
-                    if len(self.clientList)==0:
-                        print("No peers")
-                    else:
-                        print("Select a host from")
-                        print(*self.clientList,sep="\n")
-                        self.addressToSendMsg = input("host:"), int(input("port:"))
-                        if(self.addressToSendMsg not in self.clientList):
-                            print("This client is not in the client list\n")
-                        else:   
-                            self.send_message()
-                        
-                    
-                elif option == 2:
-                    self.addressToSendMsg = input("host:"), int(input("port:"))
-                    self.send_message()
-                    if(self.addressToSendMsg not in self.clientList):
-                        self.clientList.append(self.addressToSendMsg)
-                        time.sleep(0.5)
-
-                elif option==3:
-                    print('exiting')
-                    break                   
-            except:
-                print("Exception")
-                time.sleep(1)
-        reactor.stop()
-
-    def datagramReceived(self, datagram, senderAddress):
-        datagram = datagram.decode('utf-8')
-        if(datagram == "ready"):
-            reactor.callInThread(self.messageClients)
-        else:
-            print("\nMessage received....")
-            print("From", senderAddress, "::", datagram, "\n")
-            if(senderAddress not in self.clientList):
-                self.clientList.append(senderAddress)
-            time.sleep(0.5)
-
+        reactor.callInThread(self.mainMenu)
+    
+    #gets the messsage as text input, and send it to addressToSendMsg
     def send_message(self):
+        print("Sending to : ", self.addressToSendMsg)
         message = input(":::")
         self.transport.write(message.encode('utf-8'), self.addressToSendMsg)
         print(":::Message sent:::")
-        time.sleep(1)
+        self.addressLastSent = self.addressToSendMsg
         return
-
-
+    
+    #runs when a message is recieved    
+    def datagramReceived(self, datagram, senderAddress):
+        datagram = datagram.decode('utf-8')
+        
+        print("Message received....")
+        print("From", senderAddress, "::", datagram)
+        self.addNewPeer(senderAddress)
+        self.addressLastReceived = senderAddress
+        return
+        
+    def mainMenu(self):
+        while True:
+            
+            print("P2P Messanger - Menu")
+            
+            print("1 - Message Last Sent Address")
+            print("2 - Message Last Recieved Address")
+            print("3 - Message a Saved Peer")
+            print("4 - Message a New Peer")
+            print("...   ----   ...")
+            print("0 - Exit P2P Messanger")
+            print("\n\n\n")
+            
+            selection = int(input("Select an option  :  "))
+                    
+            if (selection == 1):    self.msgLastSentAddr()
+                
+            elif (selection == 2):  self.msgLastRecAddr()
+                
+            elif (selection == 3):  self.msgSavedPeer()
+                
+            elif (selection == 4):  self.msgNewPeer()
+                
+            elif (selection == 0):  self.exitApp()
+                
+            else:   print("invalid input")
+        
+            
+    def msgLastSentAddr(self):
+        self.addressToSendMsg = self.addressLastSent
+        self.send_message()
+        
+    def msgLastRecAddr(self):
+        self.addressToSendMsg = self.addressLastReceived
+        self.send_message()
+        
+    def msgSavedPeer(self):
+        self.addressToSendMsg = self.selectPeer()
+        self.send_message()
+        
+    def msgNewPeer(self):
+        self.addressToSendMsg = self.getNewPeer()
+        self.addNewPeer(self.addressToSendMsg)
+        self.send_message()
+        
+    def exitApp(self):
+        reactor.stop()
+        exit()
+    
+    def selectPeer(self):
+        print(self.peers)
+        selected = int(input("Select a peer: "))
+        return self.peers[selected]
+    
+    def getNewPeer(self):
+        peer = input("host:"), int(input("port:"))
+        return peer
+    
+    def addNewPeer(self, peer):
+        peerSize = len(self.peers)
+        #if not in the list
+        self.peers[peerSize] = peer
+        return
+    
+        
+   
 if __name__ == '__main__':
     port = randint(1000, 5000)
-    reactor.listenUDP(port, Client(port))
+    client = Client(port)
+    reactor.listenUDP(port, client)
     reactor.run()
